feat: repo tooling

+++ contracts/optimistic-ethereum/iOVM/bridge/tokens/iOVM_L1ETHGateway.sol
@@ -6,50 +6,27 @@ pragma experimental ABIEncoderV2;
  * @title iOVM_L1ETHGateway
  */
 interface iOVM_L1ETHGateway {
-
     /**********
      * Events *
      **********/
 
-    event DepositInitiated(
-        address indexed _from,
-        address _to,
-        uint256 _amount
-    );
+    event DepositInitiated(address indexed _from, address _to, uint256 _amount);
 
-    event WithdrawalFinalized(
-        address indexed _to,
-        uint256 _amount
-    );
+    event WithdrawalFinalized(address indexed _to, uint256 _amount);
 
     /********************
      * Public Functions *
      ********************/
 
-    function deposit()
-        external
-        payable;
+    function deposit() external payable;
 
-    function depositTo(
-        address _to
-    )
-        external
-        payable;
+    function depositTo(address _to) external payable;
 
     /*************************
      * Cross-chain Functions *
      *************************/
 
-    function finalizeWithdrawal(
-        address _to,
-        uint _amount
-    )
-        external;
+    function finalizeWithdrawal(address _to, uint256 _amount) external;
 
-    function getFinalizeDepositL2Gas()
-        external
-        view
-        returns(
-            uint32
-        );
+    function getFinalizeDepositL2Gas() external view returns (uint32);
 }
